Refactoring

+++ Hexagon/Util.h
@@ -6,7 +6,7 @@ using namespace sf;
 class Util
 {
 public:
-	static Vector2i pointToHex(Vector2i& pos, Vector2i offset, float dist);
+	static Vector2i pointToHex(Vector2f& pos, Vector2i offset, float dist);
 private:
 	static Vector2i roundHex(Vector2f& pos);
 };
